backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
gunicorn>=21.2
reportlab>=3.6.12
sentry-sdk>=2.13.0
psycopg2-binary>=2.9
whitenoise>=6.7
dj-database-url>=2.2
faiss-cpu>=1.7.4
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
